+++ src/customers/importers/__init__.py
@@ -1,3 +1,3 @@
-from .email_importer import CustomerImportIssue, CustomerImportResult, import_customers, import_emails
+from .email_importer import CustomerImportIssue, CustomerImportResult, apply_customer_defaults, import_customers, import_emails
 
-__all__ = ["CustomerImportIssue", "CustomerImportResult", "import_customers", "import_emails"]
+__all__ = ["CustomerImportIssue", "CustomerImportResult", "apply_customer_defaults", "import_customers", "import_emails"]